upgrading integer to float if multiplier is assigned

+++ ProtocolDrivers/ModbusTCP/ModbusTCPAsset.cs
@@ -130,7 +130,7 @@
                     {
                         value = BitConverter.ToInt32(tagBytes);
                         if (tag.Multiplier != 0.0f)
-                            value = (int)((float)value * tag.Multiplier);
+                            value = (float)value * tag.Multiplier;
                     }
                     else if (tag.Type == "String")
                     {
